[Explore] Move field selector component out of legacy folder
