+++ src/plugins/explore/public/application/legacy/discover/application/components/default_discover_table/table_header.tsx
@@ -15,7 +15,7 @@ import React from 'react';
 import { IndexPattern } from '../../../opensearch_dashboards_services';
 import { TableHeaderColumn } from './table_header_column';
 import { LegacyDisplayedColumn } from './helper';
-import { getDefaultSort } from '../../view_components/utils/get_default_sort';
+import { getDefaultSort } from '../../../../../../helpers/view_component_utils/get_default_sort';
 import { SortDirection, SortOrder } from '../../../../../../types/saved_explore_types';
 
 interface Props {
