--- src/plugins/explore/public/application/legacy/discover/application/view_components/index.ts
@@ -1,6 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './panel';
