+++ src/plugins/explore/public/components/container/bottom_container/bottom_container.tsx
@@ -7,8 +7,8 @@ import React from 'react';
 import { useSelector } from 'react-redux';
 import { EuiResizableContainer, EuiPageBody, useIsWithinBreakpoints } from '@elastic/eui';
 import { selectShowDatasetFields } from '../../../application/utils/state_management/selectors';
-import { CanvasPanel } from '../../../application/legacy/discover/application/components/panel/canvas_panel';
-import { DiscoverPanel } from '../../../application/legacy/discover/application/view_components/panel';
+import { CanvasPanel } from '../../panel/canvas_panel';
+import { DiscoverPanel } from '../../fields_selector/fields_selector_panel';
 import { BottomRightContainer } from './bottom_right_container/bottom_right_container';
 
 export const BottomContainer = () => {
