+++ src/plugins/explore/public/components/fields_selector/discover_field.test.tsx
@@ -34,8 +34,8 @@ import stubbedLogstashFields from 'fixtures/logstash_fields';
 import { render, screen, fireEvent } from 'test_utils/testing_lib_helpers';
 import { DiscoverField } from './discover_field';
 import { coreMock } from 'opensearch-dashboards/public/mocks';
-import { IndexPatternField } from '../../../../../../../../data/public';
-import { getStubIndexPattern } from '../../../../../../../../data/public/test_utils';
+import { IndexPatternField } from '../../../../data/public';
+import { getStubIndexPattern } from '../../../../data/public/test_utils';
 
 jest.mock('../../../opensearch_dashboards_services', () => ({
   getServices: () => ({
