+++ src/plugins/explore/public/components/fields_selector/discover_field.tsx
@@ -40,10 +40,10 @@ import {
 } from '@elastic/eui';
 import { i18n } from '@osd/i18n';
 import { DiscoverFieldDetails } from './discover_field_details';
-import { FieldIcon } from '../../../../../../../../opensearch_dashboards_react/public';
+import { FieldIcon } from '../../../../opensearch_dashboards_react/public';
 import { FieldDetails } from './types';
-import { IndexPatternField, IndexPattern } from '../../../../../../../../data/public';
-import { shortenDottedString } from '../../helpers';
+import { IndexPatternField, IndexPattern } from '../../../../data/public';
+import { shortenDottedString } from '../../application/legacy/discover/application/helpers';
 import { getFieldTypeName } from './lib/get_field_type_name';
 import './discover_field.scss';
