+++ src/plugins/explore/public/components/fields_selector/discover_field_search.tsx
