+++ src/plugins/explore/public/components/fields_selector/discover_sidebar.scss
