+++ src/plugins/explore/public/components/fields_selector/discover_sidebar.tsx
@@ -7,9 +7,9 @@ import { EuiButtonEmpty, EuiPanel, EuiSplitPanel } from '@elastic/eui';
 import { i18n } from '@osd/i18n';
 import { I18nProvider } from '@osd/i18n/react';
 import React, { useCallback, useMemo, useState } from 'react';
-import { OpenSearchSearchHit } from '../../../../../../types/doc_views_types';
-import { IndexPattern, IndexPatternField, UI_SETTINGS } from '../../../../../../../../data/public';
-import { getServices } from '../../../opensearch_dashboards_services';
+import { OpenSearchSearchHit } from '../../types/doc_views_types';
+import { IndexPattern, IndexPatternField, UI_SETTINGS } from '../../../../data/public';
+import { getServices } from '../../application/legacy/discover/opensearch_dashboards_services';
 import { DiscoverField } from './discover_field';
 import { DiscoverFieldSearch } from './discover_field_search';
 import './discover_sidebar.scss';
