+++ src/plugins/explore/public/components/fields_selector/index.ts
