+++ src/plugins/explore/public/components/fields_selector/lib/get_field_type_name.ts
