+++ src/plugins/explore/public/components/fields_selector/lib/group_fields.test.ts
