+++ src/plugins/explore/public/components/fields_selector/lib/group_fields.ts
@@ -28,7 +28,7 @@
  * under the License.
  */
 
-import { IndexPatternField } from '../../../../../../../../../data/public';
+import { IndexPatternField } from '../../../../../data/public';
 import { FieldFilterState, isFieldFiltered } from './field_filter';
 
 interface GroupedFields {
