+++ src/plugins/explore/public/components/fields_selector/string_progress_bar.tsx
