+++ src/plugins/explore/public/components/panel/canvas_panel.tsx
