+++ src/plugins/explore/public/helpers/view_component_utils/get_sort.test.ts
@@ -31,7 +31,7 @@
 import { getSort, getSortArray } from './get_sort';
 // @ts-ignore
 import FixturesStubbedLogstashIndexPatternProvider from 'fixtures/stubbed_logstash_index_pattern';
-import { IndexPattern } from '../../../opensearch_dashboards_services';
+import { IndexPattern } from '../../application/legacy/discover/opensearch_dashboards_services';
 
 describe('docTable', function () {
   let indexPattern: IndexPattern;
